+++ v2/bind_test.go
@@ -3,7 +3,6 @@ package v2
 import (
 	"fmt"
 	"net/http"
-	"reflect"
 	"testing"
 )
 
@@ -70,8 +69,7 @@ func TestBind(t *testing.T) {
 		expectedErr        error
 	}{
 		{
-			name:    "success - created",
-			request: defaultBindRequest(),
+			name: "success - created",
 			httpReaction: httpReaction{
 				status: http.StatusCreated,
 				body:   successBindResponseBody,
@@ -79,8 +77,7 @@ func TestBind(t *testing.T) {
 			expectedResponse: successBindResponse(),
 		},
 		{
-			name:    "success - ok",
-			request: defaultBindRequest(),
+			name: "success - ok",
 			httpReaction: httpReaction{
 				status: http.StatusOK,
 				body:   successBindResponseBody,
@@ -100,16 +97,14 @@ func TestBind(t *testing.T) {
 			expectedResponse: successBindResponse(),
 		},
 		{
-			name:    "http error",
-			request: defaultBindRequest(),
+			name: "http error",
 			httpReaction: httpReaction{
 				err: fmt.Errorf("http error"),
 			},
 			expectedErrMessage: "http error",
 		},
 		{
-			name:    "200 with malformed response",
-			request: defaultBindRequest(),
+			name: "200 with malformed response",
 			httpReaction: httpReaction{
 				status: http.StatusOK,
 				body:   malformedResponse,
@@ -117,8 +112,7 @@ func TestBind(t *testing.T) {
 			expectedErrMessage: "unexpected end of JSON input",
 		},
 		{
-			name:    "500 with malformed response",
-			request: defaultBindRequest(),
+			name: "500 with malformed response",
 			httpReaction: httpReaction{
 				status: http.StatusInternalServerError,
 				body:   malformedResponse,
@@ -126,8 +120,7 @@ func TestBind(t *testing.T) {
 			expectedErrMessage: "unexpected end of JSON input",
 		},
 		{
-			name:    "500 with conventional failure response",
-			request: defaultBindRequest(),
+			name: "500 with conventional failure response",
 			httpReaction: httpReaction{
 				status: http.StatusInternalServerError,
 				body:   conventionalFailureResponseBody,
@@ -137,6 +130,10 @@ func TestBind(t *testing.T) {
 	}
 
 	for _, tc := range cases {
+		if tc.request == nil {
+			tc.request = defaultBindRequest()
+		}
+
 		if tc.httpChecks.URL == "" {
 			tc.httpChecks.URL = "/v2/service_instances/test-instance-id/service_bindings/test-binding-id"
 		}
@@ -145,42 +142,10 @@ func TestBind(t *testing.T) {
 			tc.httpChecks.body = defaultBindRequestBody
 		}
 
-		doBindInstanceTest(t, tc.name, tc.request, tc.httpChecks, tc.httpReaction, tc.expectedResponse, tc.expectedErrMessage, tc.expectedErr)
-	}
-}
+		klient := newTestClient(t, tc.name, tc.httpChecks, tc.httpReaction)
 
-func doBindInstanceTest(
-	t *testing.T,
-	name string,
-	request *BindRequest,
-	httpChecks httpChecks,
-	httpReaction httpReaction,
-	expectedResponse *BindResponse,
-	expectedErrMessage string,
-	expectedErr error,
-) {
-	klient := &client{
-		Name:          "test client",
-		Verbose:       true,
-		URL:           "https://example.com",
-		doRequestFunc: doHTTP(t, name, httpChecks, httpReaction),
-	}
+		response, err := klient.Bind(tc.request)
 
-	response, err := klient.Bind(request)
-	if err != nil && expectedErrMessage == "" && expectedErr == nil {
-		t.Errorf("%v: error getting catalog: %v", name, err)
-		return
-	} else if err != nil && expectedErrMessage != "" && expectedErrMessage != err.Error() {
-		t.Errorf("%v: unexpected error message: expected %v, got %v", name, expectedErrMessage, err)
-		return
-	} else if err != nil && expectedErr != nil && !reflect.DeepEqual(expectedErr, err) {
-		t.Errorf("%v: unexpected error: expected %+v, got %v", name, expectedErr, err)
-		return
+		doResponseChecks(t, tc.name, response, err, tc.expectedResponse, tc.expectedErrMessage, tc.expectedErr)
 	}
-
-	if e, a := expectedResponse, response; !reflect.DeepEqual(e, a) {
-		t.Errorf("%v: unexpected diff in bind response; expected %+v, got %+v", name, e, a)
-		return
-	}
-
 }
