+++ v2/update_instance_test.go
@@ -3,7 +3,6 @@ package v2
 import (
 	"fmt"
 	"net/http"
-	"reflect"
 	"testing"
 )
 
@@ -49,8 +48,7 @@ func TestUpdateInstanceInstance(t *testing.T) {
 		expectedErr        error
 	}{
 		{
-			name:    "success - ok",
-			request: defaultUpdateInstanceRequest(),
+			name: "success - ok",
 			httpReaction: httpReaction{
 				status: http.StatusOK,
 				body:   successUpdateInstanceResponseBody,
@@ -60,6 +58,11 @@ func TestUpdateInstanceInstance(t *testing.T) {
 		{
 			name:    "success - async",
 			request: defaultAsyncUpdateInstanceRequest(),
+			httpChecks: httpChecks{
+				params: map[string]string{
+					asyncQueryParamKey: "true",
+				},
+			},
 			httpReaction: httpReaction{
 				status: http.StatusAccepted,
 				body:   successAsyncUpdateInstanceResponseBody,
@@ -69,6 +72,11 @@ func TestUpdateInstanceInstance(t *testing.T) {
 		{
 			name:    "accepted with malformed response",
 			request: defaultAsyncUpdateInstanceRequest(),
+			httpChecks: httpChecks{
+				params: map[string]string{
+					asyncQueryParamKey: "true",
+				},
+			},
 			httpReaction: httpReaction{
 				status: http.StatusAccepted,
 				body:   malformedResponse,
@@ -76,16 +84,14 @@ func TestUpdateInstanceInstance(t *testing.T) {
 			expectedErrMessage: "unexpected end of JSON input",
 		},
 		{
-			name:    "http error",
-			request: defaultUpdateInstanceRequest(),
+			name: "http error",
 			httpReaction: httpReaction{
 				err: fmt.Errorf("http error"),
 			},
 			expectedErrMessage: "http error",
 		},
 		{
-			name:    "200 with malformed response",
-			request: defaultUpdateInstanceRequest(),
+			name: "200 with malformed response",
 			httpReaction: httpReaction{
 				status: http.StatusOK,
 				body:   malformedResponse,
@@ -93,8 +99,7 @@ func TestUpdateInstanceInstance(t *testing.T) {
 			expectedResponse: successUpdateInstanceResponse(),
 		},
 		{
-			name:    "500 with malformed response",
-			request: defaultUpdateInstanceRequest(),
+			name: "500 with malformed response",
 			httpReaction: httpReaction{
 				status: http.StatusInternalServerError,
 				body:   malformedResponse,
@@ -102,8 +107,7 @@ func TestUpdateInstanceInstance(t *testing.T) {
 			expectedErrMessage: "unexpected end of JSON input",
 		},
 		{
-			name:    "500 with conventional failure response",
-			request: defaultUpdateInstanceRequest(),
+			name: "500 with conventional failure response",
 			httpReaction: httpReaction{
 				status: http.StatusInternalServerError,
 				body:   conventionalFailureResponseBody,
@@ -113,6 +117,10 @@ func TestUpdateInstanceInstance(t *testing.T) {
 	}
 
 	for _, tc := range cases {
+		if tc.request == nil {
+			tc.request = defaultUpdateInstanceRequest()
+		}
+
 		if tc.httpChecks.URL == "" {
 			tc.httpChecks.URL = "/v2/service_instances/test-instance-id"
 		}
@@ -121,41 +129,10 @@ func TestUpdateInstanceInstance(t *testing.T) {
 			tc.httpChecks.body = "{}"
 		}
 
-		doUpdateInstanceInstanceTest(t, tc.name, tc.request, tc.httpChecks, tc.httpReaction, tc.expectedResponse, tc.expectedErrMessage, tc.expectedErr)
-	}
-}
+		klient := newTestClient(t, tc.name, tc.httpChecks, tc.httpReaction)
 
-func doUpdateInstanceInstanceTest(
-	t *testing.T,
-	name string,
-	request *UpdateInstanceRequest,
-	httpChecks httpChecks,
-	httpReaction httpReaction,
-	expectedResponse *UpdateInstanceResponse,
-	expectedErrMessage string,
-	expectedErr error,
-) {
-	klient := &client{
-		Name:          "test client",
-		Verbose:       true,
-		URL:           "https://example.com",
-		doRequestFunc: doHTTP(t, name, httpChecks, httpReaction),
-	}
+		response, err := klient.UpdateInstance(tc.request)
 
-	response, err := klient.UpdateInstance(request)
-	if err != nil && expectedErrMessage == "" && expectedErr == nil {
-		t.Errorf("%v: error getting catalog: %v", name, err)
-		return
-	} else if err != nil && expectedErrMessage != "" && expectedErrMessage != err.Error() {
-		t.Errorf("%v: unexpected error message: expected %v, got %v", name, expectedErrMessage, err)
-		return
-	} else if err != nil && expectedErr != nil && !reflect.DeepEqual(expectedErr, err) {
-		t.Errorf("%v: unexpected error: expected %+v, got %v", name, expectedErr, err)
-		return
-	}
-
-	if e, a := expectedResponse, response; !reflect.DeepEqual(e, a) {
-		t.Errorf("%v: unexpected diff in catalog response; expected %+v, got %+v", name, e, a)
-		return
+		doResponseChecks(t, tc.name, response, err, tc.expectedResponse, tc.expectedErrMessage, tc.expectedErr)
 	}
 }
